refactor: move claimsmappingpolicy to serviceprincipal pkg
