+++ pkg/azure/client/serviceprincipal/claimsmappingpolicy/claimsmappingpolicy.go
