+++ pkg/azure/client/serviceprincipal/claimsmappingpolicy/validpolicy.go
